tpm: Added test tpm files for PCR extend and read
I added a tpm_test file that prints relevant information about the command.
The executable needs to be run with 'sudo'.
There is a firmware archive included in this directory with the TLCL and TSS lib
with debugging enabled. This archive can be recreated with a 'make' on the main
directory with the options DEBUG=1